+++ src/client/index.ts
@@ -1,2 +1,7 @@
 export { PushoverClient } from './pushover-client.ts'
 export { RequestBuilder } from './request-builder.ts'
+export { RetryHandler } from './retry-handler.ts'
+export { RateLimiter } from './rate-limiter.ts'
+export { MessageBuilder } from './message-builder.ts'
+export { MessageQueue } from './message-queue.ts'
+export { LimitChecker } from './limit-checker.ts'